requirements.txt
streamlit
pandas
python-dotenv
langchain
sentence-transformers
cohere
faiss-cpu
